examples/pushover_example/requirements/requirements.txt
pushover-python
python-dotenv
